+++ ui/src/components/btn-dropdown/QasBtnDropdown.vue
@@ -3,8 +3,10 @@
     <div v-if="hasButtons" :class="classes.list">
       <div v-for="(buttonProps, key, index) in props.buttonsPropsList" :key="key">
         <div class="flex no-wrap">
-          <qas-btn :disable="props.disable" v-bind="buttonProps" no-wrap variant="tertiary" @click="onClick">
-            <q-menu v-if="hasMenuOnLeftSide" v-model="isMenuOpened" anchor="bottom right" :auto-close="props.useAutoClose" class="qas-menu" self="top right" @update:model-value="onUpdateMenuValue">
+          <qas-btn v-bind="buttonProps" :data-btn-dropdown="key" :disable="props.disable" no-wrap variant="tertiary" @click="onClick">
+            <slot v-if="hasBtnContentSlot(key)" :name="`btn-content-${key}`" />
+
+            <q-menu v-else-if="hasMenuOnLeftSide" v-model="isMenuOpened" anchor="bottom right" :auto-close="props.useAutoClose" class="qas-menu" self="top right" @update:model-value="onUpdateMenuValue">
               <div :class="classes.menuContent">
                 <slot />
               </div>
@@ -141,6 +143,10 @@ function isLast (index) {
 
 function hasSeparator (index) {
   return props.useSplit || !isLast(index)
+}
+
+function hasBtnContentSlot (name) {
+  return !!slots[`btn-content-${name}`]
 }
 </script>